chore: 🔧 add dependabot configuration
For the GitHub actions and the Python backend the maintenance of
dependencies using dependabot is configured. Dependabot will run every
monday at 6 o'clock opening a pull request if necessary. The commit
message will follow the conventional commit style, also the pull request
gets labeled in order to distinguish between the updates of the
dependencies of the backup and of the github actions.