ci(backend): ♻️ use composite for shared steps
In the backend pipeline the steps to perform the code checkout,
installation of Python and the call to uv to install the dependencies
are repeately used multiple times. To reduce the code duplication these
parts are now separated into two separate yaml files. They are stored in
a separate setup-python folder. Both files only differ in the
installation of dependencies. dev.yml will install all necessary
development dependencies, while prod.yml will not. The latter can be
used e.g. to create the input for a proper software bill of materials.

The existing pipeline was adopted to use these files. Also it will now
run whenever the content in the setup-python folder gets updates.
Additionally a typo was corrected in the pipeline triggers. The file
ending of the python-backend file was yml instead of yaml.

The file ending of the yaml files should be harmonized in the future.